filter doctors

+++ shared/src/commonMain/kotlin/com/yehorsk/medical_platform_mobile/core/data/mappers/CommonMappers.kt
@@ -68,6 +68,6 @@ fun DoctorResponseDto.toDoctor() = Doctor(
 
 fun GetDoctorsWithFilter.toGetDoctorsWithFilterDto() = GetDoctorsWithFilterDto(
     search = search,
-    specializations = specializations.map { it.id.toInt() },
+    specializations = specializations.map { it.id },
     city = city
 )
